move reusable  components to StackSourceWriterHelper class
