+++ src/TraceEvent/TraceEvent.Tests/SpeedScopeExporterTests.cs
@@ -5,6 +5,8 @@ using System.Collections.Generic;
 using System.Linq;
 using Xunit;
 
+using static Microsoft.Diagnostics.Tracing.Stacks.StackSourceWriterHelper;
+
 namespace TraceEventTests
 {
     public class SpeedScopeStackSourceWriterTests
@@ -42,7 +44,7 @@ namespace TraceEventTests
 
             var stackSource = new StackSourceStub(sourceSamples);
 
-            var result = SpeedScopeStackSourceWriter.GetSortedSamplesPerThread(stackSource)[ThreadName];
+            var result = GetSortedSamplesPerThread(stackSource)[ThreadName];
 
             Assert.Equal(0.1, result[0].RelativeTime);
             Assert.Equal(0.1, result[1].RelativeTime);
@@ -75,11 +77,11 @@ namespace TraceEventTests
                 callerIndex: a.StackIndex);
 
             var allSamples = new[] { main, a, b };
-            var leafs = new[] { new SpeedScopeStackSourceWriter.Sample(b.StackIndex, -1, b.RelativeTime, b.Metric, -1) };
+            var leafs = new[] { new Sample(b.StackIndex, -1, b.RelativeTime, b.Metric, -1) };
             var stackSource = new StackSourceStub(allSamples);
             var frameNameToId = new Dictionary<string, int>();
 
-            var frameIdToSamples = SpeedScopeStackSourceWriter.WalkTheStackAndExpandSamples(stackSource, leafs, frameNameToId);
+            var frameIdToSamples = WalkTheStackAndExpandSamples(stackSource, leafs, frameNameToId);
 
             Assert.Equal(0, frameNameToId[main.Name]);
             Assert.Equal(1, frameNameToId[a.Name]);
@@ -112,11 +114,11 @@ namespace TraceEventTests
                 callerIndex: main.StackIndex);
 
             var allSamples = new[] { main, wrong };
-            var leafs = new[] { new SpeedScopeStackSourceWriter.Sample(wrong.StackIndex, -1, wrong.RelativeTime, wrong.Metric, -1) };
+            var leafs = new[] { new Sample(wrong.StackIndex, -1, wrong.RelativeTime, wrong.Metric, -1) };
             var stackSource = new StackSourceStub(allSamples);
             var frameNameToId = new Dictionary<string, int>();
 
-            var frameIdToSamples = SpeedScopeStackSourceWriter.WalkTheStackAndExpandSamples(stackSource, leafs, frameNameToId);
+            var frameIdToSamples = WalkTheStackAndExpandSamples(stackSource, leafs, frameNameToId);
 
             Assert.Equal(0, frameNameToId[main.Name]);
             Assert.False(frameNameToId.ContainsKey(wrong.Name));
@@ -133,37 +135,37 @@ namespace TraceEventTests
 
             var samples = new[]
             {
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 0.1),
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 0.2),
+                new Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 0.1),
+                new Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 0.2),
 
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 0.7),
+                new Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 0.7),
 
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 1.1),
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 1.2),
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 1.3),
+                new Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 1.1),
+                new Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 1.2),
+                new Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, depth: 0, relativeTime: 1.3),
             };
 
-            var input = new Dictionary<int, List<SpeedScopeStackSourceWriter.Sample>>() { { 0, samples.ToList() } };
+            var input = new Dictionary<int, List<Sample>>() { { 0, samples.ToList() } };
 
-            var aggregatedEvents = SpeedScopeStackSourceWriter.GetAggregatedOrderedProfileEvents(input);
+            var aggregatedEvents = GetAggregatedOrderedProfileEvents(input);
 
             // we should have <0.1, 0.3> and <0.7, 0.8> (the tool would ignore <0.7, 0.7>) and <1.1, 1.4>
             Assert.Equal(6, aggregatedEvents.Count);
 
             Assert.Equal(0.1, aggregatedEvents[0].RelativeTime);
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Open, aggregatedEvents[0].Type);
+            Assert.Equal(ProfileEventType.Open, aggregatedEvents[0].Type);
             Assert.Equal(0.2 + metric, aggregatedEvents[1].RelativeTime);
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Close, aggregatedEvents[1].Type);
+            Assert.Equal(ProfileEventType.Close, aggregatedEvents[1].Type);
 
             Assert.Equal(0.7, aggregatedEvents[2].RelativeTime);
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Open, aggregatedEvents[2].Type);
+            Assert.Equal(ProfileEventType.Open, aggregatedEvents[2].Type);
             Assert.Equal(0.7 + metric, aggregatedEvents[3].RelativeTime);
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Close, aggregatedEvents[3].Type);
+            Assert.Equal(ProfileEventType.Close, aggregatedEvents[3].Type);
 
             Assert.Equal(1.1, aggregatedEvents[4].RelativeTime);
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Open, aggregatedEvents[4].Type);
+            Assert.Equal(ProfileEventType.Open, aggregatedEvents[4].Type);
             Assert.Equal(1.3 + metric, aggregatedEvents[5].RelativeTime);
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Close, aggregatedEvents[5].Type);
+            Assert.Equal(ProfileEventType.Close, aggregatedEvents[5].Type);
         }
 
         [Fact]
@@ -173,32 +175,32 @@ namespace TraceEventTests
 
             var samples = new[]
             {
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, relativeTime: 0.1, depth: 0),
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, relativeTime: 0.2, depth: 1), // depth change!
+                new Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, relativeTime: 0.1, depth: 0),
+                new Sample((StackSourceCallStackIndex)1, callerFrameId: 0, metric: metric, relativeTime: 0.2, depth: 1), // depth change!
             };
 
-            var input = new Dictionary<int, List<SpeedScopeStackSourceWriter.Sample>>() { { 0, samples.ToList() } };
+            var input = new Dictionary<int, List<Sample>>() { { 0, samples.ToList() } };
 
-            var aggregatedEvents = SpeedScopeStackSourceWriter.GetAggregatedOrderedProfileEvents(input);
+            var aggregatedEvents = GetAggregatedOrderedProfileEvents(input);
 
             // we should have:
             //  Open at 0.1 depth 0 and Close 0.2
             //  Open at 0.2 depth 1 and Close 0.3
             Assert.Equal(4, aggregatedEvents.Count);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Open, aggregatedEvents[0].Type);
+            Assert.Equal(ProfileEventType.Open, aggregatedEvents[0].Type);
             Assert.Equal(0.1, aggregatedEvents[0].RelativeTime);
             Assert.Equal(0, aggregatedEvents[0].Depth);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Close, aggregatedEvents[1].Type);
+            Assert.Equal(ProfileEventType.Close, aggregatedEvents[1].Type);
             Assert.Equal(0.1 + metric, aggregatedEvents[1].RelativeTime);
             Assert.Equal(0, aggregatedEvents[0].Depth);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Open, aggregatedEvents[2].Type);
+            Assert.Equal(ProfileEventType.Open, aggregatedEvents[2].Type);
             Assert.Equal(0.2, aggregatedEvents[2].RelativeTime);
             Assert.Equal(1, aggregatedEvents[2].Depth);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Close, aggregatedEvents[3].Type);
+            Assert.Equal(ProfileEventType.Close, aggregatedEvents[3].Type);
             Assert.Equal(0.2 + metric, aggregatedEvents[3].RelativeTime);
             Assert.Equal(1, aggregatedEvents[3].Depth);
         }
@@ -210,32 +212,32 @@ namespace TraceEventTests
 
             var samples = new[]
             {
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, metric: metric, relativeTime: 0.1, depth: 0, callerFrameId: 0),
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, metric: metric, relativeTime: 0.2, depth: 0, callerFrameId: 1), // callerFrameId change!
+                new Sample((StackSourceCallStackIndex)1, metric: metric, relativeTime: 0.1, depth: 0, callerFrameId: 0),
+                new Sample((StackSourceCallStackIndex)1, metric: metric, relativeTime: 0.2, depth: 0, callerFrameId: 1), // callerFrameId change!
             };
 
-            var input = new Dictionary<int, List<SpeedScopeStackSourceWriter.Sample>>() { { 0, samples.ToList() } };
+            var input = new Dictionary<int, List<Sample>>() { { 0, samples.ToList() } };
 
-            var aggregatedEvents = SpeedScopeStackSourceWriter.GetAggregatedOrderedProfileEvents(input);
+            var aggregatedEvents = GetAggregatedOrderedProfileEvents(input);
 
             // we should have:
             //  Open at 0.1 depth 0 and Close 0.2
             //  Open at 0.2 depth 0 and Close 0.3
             Assert.Equal(4, aggregatedEvents.Count);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Open, aggregatedEvents[0].Type);
+            Assert.Equal(ProfileEventType.Open, aggregatedEvents[0].Type);
             Assert.Equal(0.1, aggregatedEvents[0].RelativeTime);
             Assert.Equal(0, aggregatedEvents[0].Depth);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Close, aggregatedEvents[1].Type);
+            Assert.Equal(ProfileEventType.Close, aggregatedEvents[1].Type);
             Assert.Equal(0.1 + metric, aggregatedEvents[1].RelativeTime);
             Assert.Equal(0, aggregatedEvents[0].Depth);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Open, aggregatedEvents[2].Type);
+            Assert.Equal(ProfileEventType.Open, aggregatedEvents[2].Type);
             Assert.Equal(0.2, aggregatedEvents[2].RelativeTime);
             Assert.Equal(0, aggregatedEvents[2].Depth);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Close, aggregatedEvents[3].Type);
+            Assert.Equal(ProfileEventType.Close, aggregatedEvents[3].Type);
             Assert.Equal(0.2 + metric, aggregatedEvents[3].RelativeTime);
             Assert.Equal(0, aggregatedEvents[3].Depth);
         }
@@ -247,24 +249,24 @@ namespace TraceEventTests
 
             var samples = new[]
             {
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, metric: metric, relativeTime: 0.1, depth: 0, callerFrameId: 0),
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, metric: metric, relativeTime: 0.2, depth: 0, callerFrameId: 0),
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, metric: 0.0, relativeTime: 0.3, depth: 0, callerFrameId: 0), // 0.0 metric
+                new Sample((StackSourceCallStackIndex)1, metric: metric, relativeTime: 0.1, depth: 0, callerFrameId: 0),
+                new Sample((StackSourceCallStackIndex)1, metric: metric, relativeTime: 0.2, depth: 0, callerFrameId: 0),
+                new Sample((StackSourceCallStackIndex)1, metric: 0.0, relativeTime: 0.3, depth: 0, callerFrameId: 0), // 0.0 metric
             };
 
-            var input = new Dictionary<int, List<SpeedScopeStackSourceWriter.Sample>>() { { 0, samples.ToList() } };
+            var input = new Dictionary<int, List<Sample>>() { { 0, samples.ToList() } };
 
-            var aggregatedEvents = SpeedScopeStackSourceWriter.GetAggregatedOrderedProfileEvents(input);
+            var aggregatedEvents = GetAggregatedOrderedProfileEvents(input);
 
             // we should have:
             //  Open at 0.1 depth 0 and Close 0.3
             Assert.Equal(2, aggregatedEvents.Count);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Open, aggregatedEvents[0].Type);
+            Assert.Equal(ProfileEventType.Open, aggregatedEvents[0].Type);
             Assert.Equal(0.1, aggregatedEvents[0].RelativeTime);
             Assert.Equal(0, aggregatedEvents[0].Depth);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Close, aggregatedEvents[1].Type);
+            Assert.Equal(ProfileEventType.Close, aggregatedEvents[1].Type);
             Assert.Equal(0.3, aggregatedEvents[1].RelativeTime);
             Assert.Equal(0, aggregatedEvents[0].Depth);
         }
@@ -277,58 +279,58 @@ namespace TraceEventTests
 
             var samples = new[]
             {
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, metric: zeroMetric, relativeTime: relativeTime, depth: 0, callerFrameId: 0), // 0.0 metric
-                new SpeedScopeStackSourceWriter.Sample((StackSourceCallStackIndex)1, metric: zeroMetric, relativeTime: relativeTime, depth: 0, callerFrameId: 0), // 0.0 metric and same relative time
+                new Sample((StackSourceCallStackIndex)1, metric: zeroMetric, relativeTime: relativeTime, depth: 0, callerFrameId: 0), // 0.0 metric
+                new Sample((StackSourceCallStackIndex)1, metric: zeroMetric, relativeTime: relativeTime, depth: 0, callerFrameId: 0), // 0.0 metric and same relative time
             };
 
-            var input = new Dictionary<int, List<SpeedScopeStackSourceWriter.Sample>>() { { 0, samples.ToList() } };
+            var input = new Dictionary<int, List<Sample>>() { { 0, samples.ToList() } };
 
-            Assert.Throws<ArgumentException>(() => SpeedScopeStackSourceWriter.GetAggregatedOrderedProfileEvents(input));
+            Assert.Throws<ArgumentException>(() => GetAggregatedOrderedProfileEvents(input));
         }
 
         [Fact]
         public void OrderForExportOrdersTheProfileEventsAsExpectedByTheSpeedScope()
         {
-            var profileEvents = new List<SpeedScopeStackSourceWriter.ProfileEvent>()
+            var profileEvents = new List<ProfileEvent>()
             {
-                new SpeedScopeStackSourceWriter.ProfileEvent(SpeedScopeStackSourceWriter.ProfileEventType.Open, frameId: 0, depth: 0, relativeTime: 0.1),
-                new SpeedScopeStackSourceWriter.ProfileEvent(SpeedScopeStackSourceWriter.ProfileEventType.Open, frameId: 1, depth: 1, relativeTime: 0.1),
-                new SpeedScopeStackSourceWriter.ProfileEvent(SpeedScopeStackSourceWriter.ProfileEventType.Close, frameId: 1, depth: 1, relativeTime: 0.3),
-                new SpeedScopeStackSourceWriter.ProfileEvent(SpeedScopeStackSourceWriter.ProfileEventType.Close, frameId: 0, depth: 0, relativeTime: 0.3),
-                new SpeedScopeStackSourceWriter.ProfileEvent(SpeedScopeStackSourceWriter.ProfileEventType.Open, frameId: 2, depth: 0, relativeTime: 0.3),
-                new SpeedScopeStackSourceWriter.ProfileEvent(SpeedScopeStackSourceWriter.ProfileEventType.Close, frameId: 2, depth: 0, relativeTime: 0.4),
+                new ProfileEvent(ProfileEventType.Open, frameId: 0, depth: 0, relativeTime: 0.1),
+                new ProfileEvent(ProfileEventType.Open, frameId: 1, depth: 1, relativeTime: 0.1),
+                new ProfileEvent(ProfileEventType.Close, frameId: 1, depth: 1, relativeTime: 0.3),
+                new ProfileEvent(ProfileEventType.Close, frameId: 0, depth: 0, relativeTime: 0.3),
+                new ProfileEvent(ProfileEventType.Open, frameId: 2, depth: 0, relativeTime: 0.3),
+                new ProfileEvent(ProfileEventType.Close, frameId: 2, depth: 0, relativeTime: 0.4),
             };
 
             profileEvents.Reverse(); // reverse to make sure that it does sort the elements in right way
 
-            var ordered = SpeedScopeStackSourceWriter.OrderForExport(profileEvents).ToArray();
+            var ordered = OrderForExport(profileEvents).ToArray();
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Open, ordered[0].Type);
+            Assert.Equal(ProfileEventType.Open, ordered[0].Type);
             Assert.Equal(0.1, ordered[0].RelativeTime);
             Assert.Equal(0, ordered[0].Depth);
             Assert.Equal(0, ordered[0].FrameId);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Open, ordered[1].Type);
+            Assert.Equal(ProfileEventType.Open, ordered[1].Type);
             Assert.Equal(0.1, ordered[1].RelativeTime);
             Assert.Equal(1, ordered[1].Depth);
             Assert.Equal(1, ordered[1].FrameId);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Close, ordered[2].Type);
+            Assert.Equal(ProfileEventType.Close, ordered[2].Type);
             Assert.Equal(0.3, ordered[2].RelativeTime);
             Assert.Equal(1, ordered[2].Depth);
             Assert.Equal(1, ordered[2].FrameId);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Close, ordered[3].Type);
+            Assert.Equal(ProfileEventType.Close, ordered[3].Type);
             Assert.Equal(0.3, ordered[3].RelativeTime);
             Assert.Equal(0, ordered[3].Depth);
             Assert.Equal(0, ordered[3].FrameId);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Open, ordered[4].Type);
+            Assert.Equal(ProfileEventType.Open, ordered[4].Type);
             Assert.Equal(0.3, ordered[4].RelativeTime);
             Assert.Equal(0, ordered[4].Depth);
             Assert.Equal(2, ordered[4].FrameId);
 
-            Assert.Equal(SpeedScopeStackSourceWriter.ProfileEventType.Close, ordered[5].Type);
+            Assert.Equal(ProfileEventType.Close, ordered[5].Type);
             Assert.Equal(0.4, ordered[5].RelativeTime);
             Assert.Equal(0, ordered[5].Depth);
             Assert.Equal(2, ordered[5].FrameId);
